Move CCK tests out into correct file and wrap in context block
