+++ ruby/spec/html_formatter/formatter_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'cucumber/messages'
+require 'cucumber-compatibility-kit'
 
 describe Cucumber::HTMLFormatter::Formatter do
   subject(:formatter) do
@@ -82,6 +83,36 @@ describe Cucumber::HTMLFormatter::Formatter do
       formatter.write_post_message
 
       expect(out.string).to eq("</body>\n<script>alert('Hi')</script>\n</html>")
+    end
+  end
+
+  context 'Compatibility Kit examples' do
+    Cucumber::CompatibilityKit.gherkin_examples.each do |example_name|
+      def run_formatter(messages)
+        out = StringIO.new
+        formatter = Cucumber::HTMLFormatter::Formatter.new(out)
+        formatter.process_messages(messages)
+        out.string
+      end
+
+      describe "'#{example_name}' example" do
+        subject(:html_report) { run_formatter(File.readlines(example_ndjson)) }
+
+        let(:example_ndjson) { "#{Cucumber::CompatibilityKit.example_path(example_name)}/#{example_name}.feature.ndjson" }
+
+        it { is_expected.to match(/\A<!DOCTYPE html>\s?<html/) }
+        it { is_expected.to match(/<\/html>\Z/) }
+      end
+    end
+
+    Cucumber::CompatibilityKit.markdown_examples.each do |example_name|
+      describe "'#{example_name}' example" do
+        let(:example_ndjson) { "#{Cucumber::CompatibilityKit.example_path(example_name)}/#{example_name}.feature.md.ndjson" }
+        subject(:html_report) { run_formatter(File.readlines(example_ndjson)) }
+
+        it { is_expected.to match(/\A<!DOCTYPE html>\s?<html/) }
+        it { is_expected.to match(/<\/html>\Z/) }
+      end
     end
   end
 end
